requirements.txt
accelerate==1.13.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
allpairspy==2.5.1
annotated-doc==0.0.4
anyio==4.3.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arpy==1.1.1
arrow==1.3.0
asttokens==2.4.1
async-lru==2.0.4
async-timeout==5.0.1
attrs==23.2.0
Babel==2.14.0
beautifulsoup4==4.12.3
bitarray==3.8.1
bleach==6.1.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.3.1
colorama==0.4.6
comm==0.2.2
contourpy==1.2.1
cppcheck==1.5.1
cycler==0.12.1
datasets==4.8.4
debugpy==1.8.1
decorator==5.1.1
defusedxml==0.7.1
dill==0.4.1
et_xmlfile==2.0.0
evaluate==0.4.6
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
filelock==3.25.2
fonttools==4.51.0
fqdn==1.5.1
frozenlist==1.8.0
fsspec==2026.2.0
h11==0.14.0
hf-xet==1.4.3
httpcore==1.0.5
httpx==0.27.0
huggingface_hub==1.8.0
idna==3.6
importlib_metadata==9.0.0
iniconfig==2.1.0
ipykernel==6.29.4
ipython==8.22.2
isoduration==20.11.0
jedi==0.19.1
Jinja2==3.1.3
json5==0.9.24
jsonpointer==2.4
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
jupyter-events==0.10.0
jupyter-lsp==2.2.4
jupyter_client==8.6.1
jupyter_core==5.7.2
jupyter_server==2.13.0
jupyter_server_terminals==0.5.3
jupyterlab==4.1.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.25.4
kiwisolver==1.4.5
lmdb==2.1.1
markdown-it-py==4.0.0
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.6
mdurl==0.1.2
mistune==3.0.2
mpmath==1.3.0
mulpyplexer==0.9
multidict==6.7.1
multiprocess==0.70.19
nbclient==0.10.0
nbconvert==7.16.3
nbformat==5.10.3
nest-asyncio==1.6.0
networkx==3.4.2
notebook==7.1.2
notebook_shim==0.2.4
numpy==1.26.4
openpyxl==3.1.5
overrides==7.7.0
packaging==24.0
pandas==2.2.1
pandocfilters==1.5.1
parso==0.8.3
pillow==10.3.0
platformdirs==4.2.0
pluggy==1.6.0
prometheus_client==0.20.0
prompt-toolkit==3.0.43
propcache==0.4.1
psutil==5.9.8
pure-eval==0.2.2
pyarrow==23.0.1
pycparser==3.0
pycryptodome==3.23.0
pydemumble==0.0.1
pyelftools==0.32
Pygments==2.17.2
pymannkendall==1.4.3
pyparsing==3.1.2
pypcode==3.3.3
pypdf==6.10.2
pytest==8.4.2
python-dateutil==2.9.0.post0
python-json-logger==2.0.7
pytz==2024.1
pywin32==306
pywinpty==2.0.13
pyxbe==1.0.3
pyxdia==0.1.0
PyYAML==6.0.1
pyzmq==25.1.2
referencing==0.34.0
regex==2026.3.32
requests==2.33.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==14.3.3
rpds-py==0.18.0
safetensors==0.7.0
scipy==1.15.3
Send2Trash==1.8.2
shellingham==1.5.4
six==1.16.0
smmap==5.0.3
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
stack-data==0.6.3
sympy==1.13.1
terminado==0.18.1
tibs==0.5.7
tinycss2==1.2.1
tokenizers==0.22.2
tomli==2.0.1
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tornado==6.4
tqdm==4.67.3
traitlets==5.14.2
transformers==5.5.0
typer==0.24.1
types-python-dateutil==2.9.0.20240316
typing_extensions==4.10.0
tzdata==2024.1
uri-template==1.3.0
urllib3==2.2.1
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
xxhash==3.6.0
yarl==1.23.0
z3-solver==4.13.0.0
zipp==3.23.0
